Move independent structures (curl and compression) in fetch to unexposed/internal files
